+++ src/main/java/chess/ConsoleApplication.java
@@ -2,8 +2,8 @@ package chess;
 
 import chess.domain.ChessGame;
 import chess.domain.board.Board;
-import chess.domain.board.state.BoardInitializer;
-import chess.domain.board.state.WhiteTurn;
+import chess.domain.state.BoardInitializer;
+import chess.domain.state.WhiteTurn;
 
 public class ConsoleApplication {
