+++ src/main/java/chess/domain/state/Terminate.java
@@ -1,8 +1,6 @@
-package chess.domain.board.state;
+package chess.domain.state;
 
 import chess.domain.board.Board;
-import chess.domain.board.Rank;
-import java.util.Map;
 
 public final class Terminate extends End {
