+++ src/main/java/chess/domain/state/Winner.java
@@ -1,4 +1,4 @@
-package chess.domain.board.state;
+package chess.domain.state;
 
 public enum Winner {
