+++ src/test/java/chess/domain/state/BoardInitializerTest.java
@@ -1,9 +1,10 @@
-package chess.domain.board.state;
+package chess.domain.state;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.board.Board;
 import chess.domain.piece.Piece;
+import chess.domain.state.BoardInitializer;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayName;
