+++ src/test/java/chess/domain/state/EndTest.java
@@ -1,10 +1,14 @@
-package chess.domain.board.state;
+package chess.domain.state;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.board.Board;
 import chess.domain.piece.Position;
-import java.util.HashMap;
+import chess.domain.state.BlackWin;
+import chess.domain.state.BoardInitializer;
+import chess.domain.state.GameState;
+import chess.domain.state.WhiteTurn;
+import chess.domain.state.WhiteWin;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
